VMTranslator: Only add return code once
Because the instructions from a return are independent of the function,
there's no point in including it more than once. Avoid having copies of
it by simply tracking whether we've written one yet, and once we have,
just jump to that one for all other instances of the return command.